+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 1. [](#new)
     * Added `|yaml` filter to convert input to YAML
     * Added `route` and `request` to `onPageNotFound` event
+    * Throwing exceptions from Twig templates fires `onDisplayErrorPage.[code]` event allowing better error pages
 3. [](#bugfix)
     * Fixed escaping in PageIndex::getLevelListing()
     * Fixed validation of `number` type [#3433](https://github.com/getgrav/grav/issues/3433)